Tim/feat/validate requests (#223)
* Delete parseFormFields.go

* Update controllers.go

* Update middlewares.go

* Update compare.go

* Delete sqid.go

* wip

* Update formatConnectorResponse.go

* Update formatTagResponse.go

* Update formatWorkspaceResponse.go

* Update go.sum

* Update controllers.go

* Update middlewares.go

* wip

* Update controllers.go

* wip

* validate responses

* Update controllers.go

* Update workflow-runs.go

* Update controllers.go

* Update go.mod

* Update go.sum

* Update go.mod

* Update go.sum

* Update controllers.go

* Update connectors.go

* Update formatConnectorResponse.go

* Update go.mod

* Update go.sum

* Update controllers.go

* Update controllers.go

* Update users.go

* Update controllers.go

* Update controllers.go

* Update workflows.go

* Update formatWorkflowableResponse.go

* Update go.mod

* Update go.sum

* Update credentials.go

* Update repositories.go

* Update formatAPITokenResponse.go

* Update formatRepositoryResponse.go

* Update formatWorkflowableResponse.go

* Update pagination_helpers.go

* Update workflow-runs.go

* Update repositories.go

* Update workflows.go

* wip

* wip

* Update logs.go

* Update logs.go

* Update validate.go

* Delete validate.go

* Update invites.go

* wip

* Update controllers.go

* Update go.mod

* Update go.sum

* Update go.mod

* Update go.sum

* Update repositories.go

* Update gc.go

* Update schema.go

* Update system.go

* Update formatRepositoryResponse.go

+++ irmin/db/workflow-runs.go
@@ -25,7 +25,7 @@ type WorkflowRun struct {
 
 // GetWorkflowRunsByWorkflowID returns workflow runs for the given workflow ID,
 // sorted by creation time, along with the total count of matching runs for pagination.
-func (d *Database) GetWorkflowRunsByWorkflowID(workflowID uint, limit, offset int) ([]WorkflowRun, int64, error) {
+func (d *Database) GetWorkflowRunsByWorkflowID(workflowID uint, limit, offset int) ([]WorkflowRun, int, error) {
 	// Count total number of matching events
 	var total int64
 	if err := d.Model(&WorkflowRun{}).
@@ -44,7 +44,8 @@ func (d *Database) GetWorkflowRunsByWorkflowID(workflowID uint, limit, offset in
 		Limit(limit).
 		Offset(offset).
 		Find(&workflowRuns)
-	return workflowRuns, total, result.Error
+
+	return workflowRuns, int(total), result.Error
 }
 
 func (d *Database) GetLatestWorkflowRunByWorkflowID(workflowID uint) (*WorkflowRun, error) {
